requirements.txt
gunicorn
flask
snowflake-connector-python
spacy
fuzzywuzzy